Move LLM mocks into individual test files instead of shared conftest
Each test file that calls optimize_anything (test_refiner.py,
test_cache_evaluation_storage.py, test_best_example_evals.py) now has
its own autouse _mock_litellm fixture. No changes to shared conftest
or pyproject.toml — keeps the mock scoped to only the files that need it.
